requirements.txt
typer[all]
rich
sqlalchemy
requests
pandas
folium
python-dotenv
newspaper3k
feedparser
duckduckgo-search
beautifulsoup4
lxml
pystac-client
geopy
python-dateutil